updated models

+++ src/main/java/com/example/boardmanager/repository/ColumnRepository.java
@@ -1,8 +1,7 @@
 package com.example.boardmanager.repository;
 
 import com.example.boardmanager.model.Column;
-import java.math.BigInteger;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ColumnRepository extends JpaRepository<Column, BigInteger> {
+public interface ColumnRepository extends JpaRepository<Column, Long> {
 }
